v0.1.0

+++ src/main/java/de/uniwue/helper/RecognitionHelper.java
@@ -1,6 +1,7 @@
 package de.uniwue.helper;
 
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.FileVisitOption;
 import java.nio.file.Files;
@@ -9,8 +10,10 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.TreeMap;
 
 import org.apache.commons.io.FileUtils;
@@ -42,6 +45,12 @@ public class RecognitionHelper {
     private String projectImageType;
 
     /**
+     * Processing structure of the project
+     * Possible values: { Directory, Pagexml }
+     */
+    private String processingMode;
+
+    /**
      * Object to use generic functionalities
      */
     private GenericHelper genericHelper;
@@ -65,6 +74,11 @@ public class RecognitionHelper {
      * Indicates if a Recognition process is already running
      */
     private boolean RecognitionRunning = false;
+
+    /**
+     * Last time the images/pagexml are modified
+     */
+    private Map<String,Long> imagesLastModified;
 
     /**
      * Structure to monitor the progress of the process
@@ -91,12 +105,15 @@ public class RecognitionHelper {
      *
      * @param projectDir Path to the project directory
      * @param projectImageType Type of the project (binary, gray)
+     * @param processingMode Processing structure of the project (Directory, Pagexml)
+     * 
      */
-    public RecognitionHelper(String projectDir, String projectImageType) {
+    public RecognitionHelper(String projectDir, String projectImageType, String processingMode) {
         this.projectImageType = projectImageType;
+        this.processingMode = processingMode;
         projConf = new ProjectConfiguration(projectDir);
         genericHelper = new GenericHelper(projConf);
-        procStateCol = new ProcessStateCollector(projConf, projectImageType);
+        procStateCol = new ProcessStateCollector(projConf, projectImageType, processingMode);
         processHandler = new ProcessHandler();
     }
 
@@ -116,34 +133,43 @@ public class RecognitionHelper {
      * @throws IOException
      */
     public void initialize(List<String> pageIds) throws IOException {
-        // Initialize the status structure
-        processState = new TreeMap<String, TreeMap<String, TreeMap<String, Boolean>>>();
+    	if(processingMode.equals("Directory")) {
+			// Initialize the status structure
+			processState = new TreeMap<String, TreeMap<String, TreeMap<String, Boolean>>>();
 
-        for (String pageId : pageIds) {
-            TreeMap<String, TreeMap<String, Boolean>> segments = new TreeMap<String, TreeMap<String, Boolean>>();
-            // File depth of 1 -> no recursive (file)listing
-            File[] lineSegmentDirectories = new File(projConf.PAGE_DIR + pageId).listFiles(File::isDirectory);
-            if (lineSegmentDirectories.length != 0) {
-                for (File dir : lineSegmentDirectories) {
-                    TreeMap<String, Boolean> lineSegments = new TreeMap<String, Boolean>();
-                    Files.walk(Paths.get(dir.getAbsolutePath()), 1)
-                    .map(Path::toFile)
-                    .filter(fileEntry -> fileEntry.isFile())
-                    .filter(fileEntry -> fileEntry.getName().endsWith(projConf.getImageExtensionByType(projectImageType)))
-                    .forEach(
-                        fileEntry -> {
-                            // Line segments have one of the following endings: ".bin.png" | ".nrm.png"
-                            // Therefore both extensions need to be removed
-                            String lineSegmentId = FilenameUtils.removeExtension(FilenameUtils.removeExtension(fileEntry.getName()));
-                            lineSegments.put(lineSegmentId, false);
-                        }
-                    );
-                    segments.put(dir.getName(), lineSegments);
-                }
-            }
+			for (String pageId : pageIds) {
+				TreeMap<String, TreeMap<String, Boolean>> segments = new TreeMap<String, TreeMap<String, Boolean>>();
+				// File depth of 1 -> no recursive (file)listing
+				File[] lineSegmentDirectories = new File(projConf.PAGE_DIR + pageId).listFiles(File::isDirectory);
+				if (lineSegmentDirectories.length != 0) {
+					for (File dir : lineSegmentDirectories) {
+						TreeMap<String, Boolean> lineSegments = new TreeMap<String, Boolean>();
+						Files.walk(Paths.get(dir.getAbsolutePath()), 1)
+						.map(Path::toFile)
+						.filter(fileEntry -> fileEntry.isFile())
+						.filter(fileEntry -> fileEntry.getName().endsWith(projConf.getImageExtensionByType(projectImageType)))
+						.forEach(
+							fileEntry -> {
+								// Line segments have one of the following endings: ".bin.png" | ".nrm.png"
+								// Therefore both extensions need to be removed
+								String lineSegmentId = FilenameUtils.removeExtension(FilenameUtils.removeExtension(fileEntry.getName()));
+								lineSegments.put(lineSegmentId, false);
+							}
+						);
+						segments.put(dir.getName(), lineSegments);
+					}
+				}
 
-            processState.put(pageId, segments);
-        }
+				processState.put(pageId, segments);
+			}
+    	} else {
+			// Init the listener for image modification
+			imagesLastModified = new HashMap<>();
+			for(String pageId: pageIds) {
+				final String pageXML = projConf.OCR_DIR + pageId + projConf.CONF_EXT;
+				imagesLastModified.put(pageXML,new File(pageXML).lastModified());
+			}
+    	}
     }
 
     /**
@@ -177,27 +203,42 @@ public class RecognitionHelper {
         if (RecognitionRunning == false)
             return progress;
 
-        int lineSegmentCount = 0;
-        int processedLineSegmentCount = 0;
-         // Identify how many line segments are already processed
-        for (String pageId : processState.keySet()) {
-            for (String segmentId : processState.get(pageId).keySet()) {
-                for (String lineSegmentId : processState.get(pageId).get(segmentId).keySet()) {
-                    lineSegmentCount += 1;
+		if(processingMode.equals("Directory")) {
+			int lineSegmentCount = 0;
+			int processedLineSegmentCount = 0;
+			 // Identify how many line segments are already processed
+			for (String pageId : processState.keySet()) {
+					for (String segmentId : processState.get(pageId).keySet()) {
+						for (String lineSegmentId : processState.get(pageId).get(segmentId).keySet()) {
+							lineSegmentCount += 1;
 
-                    if(processState.get(pageId).get(segmentId).get(lineSegmentId)) {
-                        processedLineSegmentCount += 1;
-                        continue;
-                    }
+							if(processState.get(pageId).get(segmentId).get(lineSegmentId)) {
+								processedLineSegmentCount += 1;
+								continue;
+							}
 
-                    if (new File(projConf.PAGE_DIR + pageId + File.separator + segmentId +
-                            File.separator + lineSegmentId + projConf.REC_EXT).exists()) {
-                        processState.get(pageId).get(segmentId).put(lineSegmentId, true);
-                    }
-                }
-            }
-        }
-        return (progress != 100) ? (int) ((double)processedLineSegmentCount / lineSegmentCount * 100) : 100;
+							if (new File(projConf.PAGE_DIR + pageId + File.separator + segmentId +
+									File.separator + lineSegmentId + projConf.REC_EXT).exists()) {
+								processState.get(pageId).get(segmentId).put(lineSegmentId, true);
+							}
+						}
+					}
+			}
+			return (progress != 100) ? (int) ((double)processedLineSegmentCount / lineSegmentCount * 100) : 100;
+		} else {
+			int modifiedCount = 0;
+			if(imagesLastModified != null) {
+				for(String pagexml : imagesLastModified.keySet()) {
+					if(imagesLastModified.get(pagexml) < new File(pagexml).lastModified()) {
+						modifiedCount++;
+					}
+				}
+				progress = (modifiedCount*100) / imagesLastModified.size();
+			} else {
+				progress = -1;
+			}
+			return progress;
+		}
     }
 
     /**
@@ -238,9 +279,9 @@ public class RecognitionHelper {
             }
         }
 
-        // Reset recognition data
-        deleteOldFiles(pageIds);
-        initialize(pageIds);
+		// Reset recognition data
+		deleteOldFiles(pageIds);
+		initialize(pageIds);
 
         List<String> command = new ArrayList<String>();
         List<String> lineSegmentImages = getLineSegmentImagesForCurrentProcess(pageIds);
@@ -251,9 +292,17 @@ public class RecognitionHelper {
         segmentListFile.deleteOnExit(); // Delete if OCR4all terminates
         ObjectMapper mapper = new ObjectMapper();
         ArrayNode segmentList = mapper.createArrayNode();
-        for (String lineSegmentImage : lineSegmentImages) {
-            // Add affected line segment images with their absolute path to the json file
-        	segmentList.add(lineSegmentImage);
+        if(processingMode.equals("Pagexml")) {
+			for (String pageId : pageIds) {
+				// Add affected images with their absolute path to the json file
+				segmentList.add(projConf.getImageDirectoryByType(projectImageType) + pageId + 
+									projConf.getImageExtensionByType(projectImageType));
+			}
+        } else {
+			for (String lineSegmentImage : lineSegmentImages) {
+				// Add affected line segment images with their absolute path to the json file
+				segmentList.add(lineSegmentImage);
+			}
         }
         ObjectNode segmentObj = mapper.createObjectNode();
         segmentObj.set("files", segmentList);
@@ -274,6 +323,15 @@ public class RecognitionHelper {
 
         command.add("--no_progress_bars");
 
+        if(processingMode.equals("Pagexml")) {
+        	command.add("--dataset");
+        	command.add("PAGEXML");
+        	// Set output extension to input extension in order to overwrite the original file
+        	// (default would've been .pred.xml)
+        	command.add("--extension");
+        	command.add(".xml");
+        }
+
         processHandler = new ProcessHandler();
         processHandler.setFetchProcessConsole(true);
         processHandler.startProcess("calamari-predict", command, false);
@@ -281,7 +339,9 @@ public class RecognitionHelper {
         // Execute progress update to fill processState data structure with correct values
         getProgress();
         // Process extension to ocropus-gpageseg script
-        createSkippedSegments();
+        if(processingMode.equals("Pagexml")) {
+			createSkippedSegments();
+        }
 
         progress = 100;
         RecognitionRunning = false;
@@ -331,31 +391,50 @@ public class RecognitionHelper {
      *
      * @param pageIds Identifiers of the pages (e.g 0002,0003)
      */
-    public void deleteOldFiles(List<String> pageIds) {
-        // Delete all files created by subsequent processes to preserve data integrity
-        ResultGenerationHelper resultGenerationHelper = new ResultGenerationHelper(projConf.PROJECT_DIR, projectImageType);
-        resultGenerationHelper.deleteOldFiles(pageIds, "txt");
-        resultGenerationHelper.deleteOldFiles(pageIds, "xml");
+    public void deleteOldFiles(List<String> pageIds) throws IOException {
+    	if(processingMode.equals("Directory")) {
+			// Delete all files created by subsequent processes to preserve data integrity
+			ResultGenerationHelper resultGenerationHelper = new ResultGenerationHelper(projConf.PROJECT_DIR, projectImageType, processingMode);
+			resultGenerationHelper.deleteOldFiles(pageIds, "txt");
 
-        for(String pageId : pageIds) {
-            File pageDirectory = new File(projConf.PAGE_DIR + pageId);
-            if (!pageDirectory.exists())
-                return;
+			for(String pageId : pageIds) {
+				File pageDirectory = new File(projConf.PAGE_DIR + pageId);
+				if (!pageDirectory.exists())
+					return;
 
-            File[] lineSegmentDirectories = pageDirectory.listFiles(File::isDirectory);
-            if (lineSegmentDirectories.length != 0) {
-                for (File dir : lineSegmentDirectories) {
-                    // Delete .txt files that store the recognized text
-                    // Keep .gt.txt files that store already manually corrected text
-                    File[] txtFiles = new File(dir.getAbsolutePath()).listFiles(
-                        (d, name) -> name.endsWith(projConf.REC_EXT) && !name.endsWith(projConf.GT_EXT)
-                    );
-                    for (File txtFile : txtFiles) {
-                        txtFile.delete();
-                    }
-                }
-            }
-        }
+				File[] lineSegmentDirectories = pageDirectory.listFiles(File::isDirectory);
+				if (lineSegmentDirectories.length != 0) {
+					for (File dir : lineSegmentDirectories) {
+						// Delete .txt files that store the recognized text
+						// Keep .gt.txt files that store already manually corrected text
+						File[] txtFiles = new File(dir.getAbsolutePath()).listFiles(
+							(d, name) -> name.endsWith(projConf.REC_EXT) && !name.endsWith(projConf.GT_EXT)
+						);
+						for (File txtFile : txtFiles) {
+							txtFile.delete();
+						}
+					}
+				}
+			}
+    	} else {
+    		// Delete potential TextEquivs already existing in the page xmls
+			for(String pageId : pageIds) {
+				File pageXML = new File(projConf.OCR_DIR + pageId + projConf.CONF_EXT);
+				if (!pageXML.exists())
+					return;
+			   
+				// Load pageXML and replace/delete all Textline text content
+				String pageXMLContent = new String(Files.readAllBytes(pageXML.toPath()));
+				pageXMLContent = pageXMLContent.replaceAll("\\<TextEquiv[^>]+?index=\"[^0]\"[^>]*?\\>[^<]*?\\<\\/TextEquiv\\>", "");
+				
+				// Save new pageXML
+				try (FileWriter fileWriter = new FileWriter(pageXML)) {
+					fileWriter.write(pageXMLContent);
+					fileWriter.flush();
+					fileWriter.close();
+				}
+			}
+    	}
     }
 
     /**
